+++ numerapi/utils.py
@@ -6,3 +6,18 @@ def parse_datetime_string(s):
         return None
     dt = dateutil.parser.parse(s)
     return dt
+
+
+def parse_float_string(s):
+    if s is None:
+        return None
+    try:
+        f = float(s.replace(",", ""))
+    except ValueError:
+        f = None
+    return f
+
+
+def replace(dictionary, key, function):
+    if dictionary is not None and key in dictionary:
+        dictionary[key] = function(dictionary[key])